modulo de noticias e imagenes terminado

+++ src/BackendBundle/Controller/NoticeController.php
@@ -3,31 +3,38 @@
 namespace BackendBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use BackendBundle\Entity\Notice;
 
 class NoticeController extends Controller
 {
+
+    //---------------------INDEX------------------------------
+
     public function indexAction(){
         
         $con = $this->getDoctrine()->getManager();
         $notice = $con->getRepository('BackendBundle:Notice')->findAll();
 
-        return $this->render('notice/index.html.twig', array('notice' => $notice));
+        $delete_form = $this->createCustomForm(':ID','DELETE','notice_delete');
+
+        return $this->render('notice/index.html.twig', array('notice' => $notice,'delete_form' => $delete_form->createView() ));
 
     }
 
+    //---------------------EDIT------------------------------
+
     public function editAction(Request $request, Notice $notice){
         
-        $deleteForm = $this->createDeleteForm($notice);
-
         $editForm = $this->createForm('BackendBundle\Form\NoticeType', $notice);
         $editForm->handleRequest($request);
 
         if ($editForm->isSubmitted() && $editForm->isValid()) {
 
             if(!empty($notice->fileIds)){
-                $notice->setImagen($notice->fileIds);
+                $notice->setImgSrc("uploads/notice/".$notice->fileIds);
             }
 
             $this->getDoctrine()->getManager()->flush();
@@ -38,10 +45,11 @@ class NoticeController extends Controller
         return $this->render('notice/edit.html.twig', array(
             'notice' => $notice,
             'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
         ));
 
     }
+
+    //---------------------NEW------------------------------
 
     public function newAction(Request $request)
     {
@@ -52,7 +60,7 @@ class NoticeController extends Controller
         if ($form->isSubmitted() && $form->isValid()) {
 
             if(!empty($notice->fileIds)){
-                $notice->setImgSrc($notice->fileIds);
+                $notice->setImgSrc("uploads/notice/".$notice->fileIds);
             }
             
             $em = $this->getDoctrine()->getManager();
@@ -68,26 +76,35 @@ class NoticeController extends Controller
         ));
     }
 
-    public function deleteAction(Request $request, Notice $notice)
+    //---------------------DELETE------------------------------
+
+    public function deleteAction(Request $request, $id)
     {
-        $form = $this->createDeleteForm($notice);
+        $em = $this->getDoctrine()->getManager();
+        $notice = $em->getRepository('BackendBundle:Notice')->find($id);
+
+        $form = $this->createCustomForm($notice->getId(),'DELETE', 'notice_delete');
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($notice);
-            $em->flush();
-        }
+        if($form->isSubmitted() && $form->isValid()){
+            
+            if($request->isXMLHttpRequest()){
+                
+                $em->remove($notice) ;
+                $em->flush();   
+                return new Response(json_encode(array('removed' => 1,'message' => 'Noticia borrado')),200, array('Content-Type' => 'application/json'));
+            }
 
-        return $this->redirectToRoute('notice_index');
+            $em->remove($notice) ;
+            $em->flush();
+
+            return $this->redirectToRoute('notice_index');
+        }
     }
 
-    private function createDeleteForm(Notice $notice)
-    {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('notice', array('id' => $notice->getId())))
-            ->setMethod('DELETE')
-            ->getForm()
-        ;
+    //---------------------FORMS------------------------------
+
+    private function createCustomForm($id,$method,$route){
+        return $this->createFormBuilder()->setAction($this->generateUrl($route, array ('id' => $id)))->setMethod($method)->getForm();
     }
 }
